修HTS介面
